+++ client/src/App.js
@@ -4,6 +4,7 @@ import "./components/style.css";
 import NewUser from "./components/NewUser/newuser";
 import { Route, BrowserRouter as Router } from "react-router-dom";
 import Profile from "./pages/Profile/profile";
+import Prediction from "./pages/Prediction/Prediction"
 
 
 class App extends Component {
@@ -13,6 +14,7 @@ class App extends Component {
         <Route exact path="/" component={LoginForm} />
         <Route exact path="/newuser" component={NewUser} />
         <Route exact path="/profile" component={Profile} />
+        <Route exact path="/prediction" component={Prediction} />
       </Router>
     );
   }
