+++ package.json
@@ -31,6 +31,7 @@
     "express-session": "^1.16.1",
     "if-env": "^1.0.4",
     "mysql2": "^1.6.5",
+    "node-fetch": "^2.6.0",
     "passport": "^0.4.0",
     "react-router-dom": "^5.0.0",
     "sequelize": "^5.8.5"
